modified csv files & updated sql script
I'm having issues with the last two tables "volcano" and "wildfire", something wrong with the format.  I had to remove the index column from the csv files, that worked to fix "fireball" "storm" and "ice", but didn't do the trick for the last two.  Any suggestions? Anyone want to try to get the data to import?  We do have more than enough records from the combined tables.